Refactored to SetGameView logic to the controller in preperation for animation
